requirements.txt
aiofile==3.8.8
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blis==0.7.11
caio==0.9.17
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.5
configobj==5.0.8
configparser==7.0.0
cymem==2.0.8
dataclasses-json==0.6.6
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
etelemetry==0.3.1
fastapi==0.111.0
fastapi-cli
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.5.0
grpcio==1.64.0
grpcio-tools==1.64.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
hyperframe==6.0.1
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.2
langchain-openai==0.1.8
langchain-qdrant==0.1.0
langchain-text-splitters==0.2.0
langcodes==3.4.0
langsmith==0.1.64
language_data==1.2.0
looseversion==1.3.0
lxml==5.2.2
marisa-trie==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
openai==1.30.5
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pathlib==1.0.1
portalocker==2.8.2
preshed==3.0.9
protobuf==5.27.0
prov==2.0.0
pydantic==2.7.2
pydantic_core==2.18.3
pydot==2.0.0
Pygments==2.18.0
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
qdrant-client==1.9.1
rdflib==7.0.0
regex==2024.5.15
requests==2.32.3
rich==13.7.1
scipy==1.13.1
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
SQLAlchemy==2.0.30
srsly==2.4.8
starlette==0.37.2
tenacity==8.3.0
thinc==8.2.4
tiktoken==0.7.0
tqdm==4.66.4
traits==6.3.2
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
